Tests/LibWeb/Text/expected/UIEvents/mouse-button-events.txt

> single left click
pointerdown target=#target button=0 detail=0
mousedown target=#target button=0 detail=0
pointerup target=#target button=0 detail=0
mouseup target=#target button=0 detail=0
click target=#target button=0 detail=0
> double left click
dblclick target=#target button=0 detail=0
pointerup target=#target button=0 detail=0
mouseup target=#target button=0 detail=0
click target=#target button=0 detail=0
> triple left click
pointerup target=#target button=0 detail=0
mouseup target=#target button=0 detail=0
click target=#target button=0 detail=0
> single right click
pointerdown target=#target button=2 detail=0
mousedown target=#target button=2 detail=0
pointerup target=#target button=2 detail=0
mouseup target=#target button=2 detail=0
contextmenu target=#target button=2 detail=0
> single right click with contextmenu default behavior prevented
pointerdown target=#target button=2 detail=0
mousedown target=#target button=2 detail=0
pointerup target=#target button=2 detail=0
mouseup target=#target button=2 detail=0
contextmenu target=#target button=2 detail=0
